Move econet base entity to separate module
