+++ homeassistant/components/econet/__init__.py
@@ -16,14 +16,12 @@ from pyeconet.errors import (
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_EMAIL, CONF_PASSWORD, Platform
-from homeassistant.core import HomeAssistant, callback
+from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
-from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.dispatcher import async_dispatcher_connect, dispatcher_send
-from homeassistant.helpers.entity import Entity
+from homeassistant.helpers.dispatcher import dispatcher_send
 from homeassistant.helpers.event import async_track_time_interval
 
-from .const import API_CLIENT, DOMAIN, EQUIPMENT
+from .const import API_CLIENT, DOMAIN, EQUIPMENT, PUSH_UPDATE
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -34,7 +32,6 @@ PLATFORMS = [
     Platform.SWITCH,
     Platform.WATER_HEATER,
 ]
-PUSH_UPDATE = "econet.push_update"
 
 INTERVAL = timedelta(minutes=60)
 
@@ -99,41 +96,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.data[DOMAIN][API_CLIENT].pop(entry.entry_id)
         hass.data[DOMAIN][EQUIPMENT].pop(entry.entry_id)
     return unload_ok
-
-
-class EcoNetEntity(Entity):
-    """Define a base EcoNet entity."""
-
-    _attr_should_poll = False
-
-    def __init__(self, econet):
-        """Initialize."""
-        self._econet = econet
-        self._attr_name = econet.device_name
-        self._attr_unique_id = f"{econet.device_id}_{econet.device_name}"
-
-    async def async_added_to_hass(self):
-        """Subscribe to device events."""
-        await super().async_added_to_hass()
-        self.async_on_remove(
-            async_dispatcher_connect(self.hass, PUSH_UPDATE, self.on_update_received)
-        )
-
-    @callback
-    def on_update_received(self):
-        """Update was pushed from the ecoent API."""
-        self.async_write_ha_state()
-
-    @property
-    def available(self):
-        """Return if the device is online or not."""
-        return self._econet.connected
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return device registry information for this entity."""
-        return DeviceInfo(
-            identifiers={(DOMAIN, self._econet.device_id)},
-            manufacturer="Rheem",
-            name=self._econet.device_name,
-        )
