+++ homeassistant/components/econet/const.py
@@ -3,3 +3,5 @@
 DOMAIN = "econet"
 API_CLIENT = "api_client"
 EQUIPMENT = "equipment"
+
+PUSH_UPDATE = "econet.push_update"
