+++ homeassistant/components/econet/switch.py
@@ -13,8 +13,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import EcoNetEntity
 from .const import DOMAIN, EQUIPMENT
+from .entity import EcoNetEntity
 
 _LOGGER = logging.getLogger(__name__)
